Add Migration, update DiamondPrice, diamond criteria, update diamonds, add little fields to order

+++ DiamondShop.Application/Usecases/Diamonds/Commands/Delete/DeleteDiamondCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using DiamondShop.Application.Services.Interfaces;
 using DiamondShop.Commons;
+using DiamondShop.Domain.Common.Enums;
 using DiamondShop.Domain.Models.Diamonds.ValueObjects;
 using DiamondShop.Domain.Repositories;
 using FluentResults;
@@ -30,10 +31,12 @@ namespace DiamondShop.Application.Usecases.Diamonds.Commands.Delete
             var getDiamond = await _diamondRepository.GetById(diamondId);
             if (getDiamond == null)
                 return Result.Fail(new NotFoundError("not found this diamond"));
-            //if (getDiamond.Warranty is not null)
-            //    return Result.Fail(new ConflictError("this diamond seems to have a warranty set to it, so it is bought already, cannot delelte"));
-            if(getDiamond.JewelryId is not null)
+            if(getDiamond.Status == ProductStatus.Sold)
+                return Result.Fail(new ConflictError("cannot delete a sold diamond"));
+            if (getDiamond.JewelryId is not null)
                 return Result.Fail(new ConflictError("this diamond seems to have a jewelry set to it, so it is used already, cannot delelte"));
+            if (getDiamond.Status != ProductStatus.Inactive)
+                return Result.Fail(new ConflictError("cannot delete a diamond that is not inactive, try to set it to inactive first "));
             await _diamondRepository.Delete(getDiamond);
             await _unitOfWork.SaveChangesAsync();
             return Result.Ok();
